// src/app/models/matricula.ts
export interface Matricula {
  id: number;
  alumnoId: number;
  codigoAlumno: string;
  alumnoNombreCompleto: string;
  gradoId: number;
  grado: string;
  nivel: string;
  seccionId: number;
  seccion: string;
  periodoAcademicoId: number;
  periodoAcademico: string;
  anioAcademico: number;
  fechaMatricula: string;
}